+++ package.json
@@ -11,7 +11,8 @@
     "react-bootstrap": "^1.0.0-beta.16",
     "react-dom": "^16.12.0",
     "react-scripts": "3.2.0",
-    "typescript": "3.7.2"
+    "typescript": "3.7.2",
+    "unirest": "^0.6.0"
   },
   "scripts": {
     "start": "react-scripts start",
